+++ CHANGES.rst
@@ -53,6 +53,8 @@ Unreleased
 - Remove ``WebSocketException.original_exception`` method.
 - Remove the ``after`` argument for :meth:`.conversations`.
 - Remove key ``reset_timestamp`` from :meth:`.limits`.
+- Remove ``SubredditMessage.mute`` and ``SubredditMessage.unmute`` methods.
+- Remove ``InboxableMixin.unblock_subreddit`` method.
 
 7.8.1 (2024/10/25)
 ------------------
@@ -213,7 +215,7 @@ Unreleased
   WikiPage.
 - :meth:`.revert` to revert a WikiPage to a specified revision.
 - :meth:`.Inbox.mark_all_read` to mark all messages as read with one API call.
-- :meth:`~.InboxableMixin.unblock_subreddit` to unblock a subreddit.
+- ``InboxableMixin.unblock_subreddit`` to unblock a subreddit.
 - :meth:`.update_crowd_control_level` to update the crowd control level of a post.
 - :meth:`.moderator_subreddits`, which returns information about the subreddits that the
   authenticated user moderates, has been restored.